+++ frontend/src/components/Dashboard/TopNav.vue
@@ -65,7 +65,3 @@
     }
 }
 </script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-</style>