+++ frontend/src/pages/Index.vue
@@ -1,7 +1,16 @@
 <template>
     <div class="container col-md-7 col-sm-12 sm-container">
 
-        <Header/>
+      <Header/>
+
+      <div v-if="!loaded" class="row mt-5 mb-5">
+        <div class="col-12 mt-5 mb-2 text-center">
+          <font-awesome-icon icon="circle-notch" class="text-dim" size="3x" spin/>
+        </div>
+        <div class="col-12 text-center mt-3 mb-3">
+          <span class="text-dim">{{loading_text}}</span>
+        </div>
+      </div>
 
         <div class="col-12 full-col-12">
             <div v-for="service in services_no_group" v-bind:key="service.id" class="list-group online_list mb-4">
@@ -14,9 +23,9 @@
             </div>
         </div>
 
-        <div>
-            <Group v-for="group in groups" v-bind:key="group.id" :group=group />
-        </div>
+      <div v-if="loaded">
+        <Group v-for="group in groups" v-bind:key="group.id" :group=group />
+      </div>
 
         <div class="col-12 full-col-12">
             <MessageBlock v-for="message in messages" v-bind:key="message.id" :message="message" />
@@ -24,7 +33,7 @@
 
         <div class="col-12 full-col-12">
             <div v-for="service in services" :ref="service.id" v-bind:key="service.id">
-                <ServiceBlock :in_service=service />
+                <ServiceBlock :service="service" />
             </div>
         </div>
 
@@ -52,10 +61,29 @@ export default {
     },
     data() {
         return {
-            logged_in: false
+            logged_in: false,
         }
     },
     computed: {
+      loading_text() {
+        if (this.core == null) {
+          return "Loading Core"
+        } else if (this.groups == null) {
+          return "Loading Groups"
+        } else if (this.services == null) {
+          return "Loading Services"
+        } else if (this.messages == null) {
+          return "Loading Announcements"
+        } else {
+          return "Completed"
+        }
+      },
+      loaded() {
+        return this.core !== null && this.groups !== null && this.services !== null
+      },
+        core() {
+          return this.$store.getters.core
+        },
         messages() {
             return this.$store.getters.messages.filter(m => this.inRange(m) && m.service === 0)
         },