+++ frontend/vue.config.js
@@ -2,6 +2,9 @@ module.exports = {
   baseUrl: '/',
   assetsDir: 'assets',
   filenameHashing: false,
+  productionTip: process.env.NODE_ENV !== 'production',
+  devtools: process.env.NODE_ENV !== 'production',
+  performance: process.env.NODE_ENV !== 'production',
   devServer: {
     disableHostCheck: true,
     proxyTable: {