+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added Help page that is generated from Statping's Wiki repo on build
 - Modified Service Group failures on index page to show 90 days of failures
 - Modified Service view page, updated Latency and Ping charts, added failures below
+- Modified Service chart on index page to show ping data along with latency
 
 # 0.90.63 (08-17-2020)
 - Modified build process to use xgo for all arch builds